[#280]feat: 코드 실행 버튼 RightSection으로 이동
