+++ src/pages/Visualization/components/LeftSection/LeftSection.module.css
@@ -15,21 +15,6 @@
   align-items: center;
   gap: 16px;
 }
-.playcode-btn {
-  background-color: var(--blue);
-  color: var(--white);
-  width: 112px;
-  height: 35px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 4px;
-  flex-direction: row;
-  border-radius: 4px;
-}
-.playcode-btn-loading {
-  filter: brightness(0.9);
-}
 .select-box {
   position: relative;
   width: 193px;
