+++ src/pages/Visualization/components/LeftSection/LeftSection.tsx
@@ -5,13 +5,7 @@ import styles from "./LeftSection.module.css";
 import CodeEditor from "./components/CodeEditor";
 import Console from "./components/Console";
 import Split from "react-split";
-import { runCode } from "@/services/api";
-import { useMutation } from "@tanstack/react-query";
 import Dropdown from "./components/Dropdown";
-import { useConsoleStore, useCodeFlowLengthStore } from "@/store/console";
-import { useEditorStore } from "@/store/editor";
-
-import { useCustomAlert } from "@/pages/components/CustomAlert.tsx";
 
 
 import { PreprocessedCodesContext } from "../../context/PreProcessedCodesContext";
@@ -24,56 +18,11 @@ const LeftSection = () => {
   if (!preprocessedCodesContext) {
     throw new Error("preprocessedCodesContext not found"); //context가 없을 경우 에러 출력 패턴 처리안해주면 에러 발생
   }
-  const setErrorLine = useEditorStore((state) => state.setErrorLine);
-  const setConsoleList = useConsoleStore((state) => state.setConsoleList);
-  const setStepIdx = useConsoleStore((state) => state.setStepIdx);
-  const { setPreprocessedCodes } = preprocessedCodesContext;
-  const setCodeFlowLength = useCodeFlowLengthStore((state) => state.setCodeFlowLength);
-  const setHighlightLines = useEditorStore((state) => state.setHighlightLines);
-
-  const { inputData } = useConsoleStore();
   const codeContext = useContext(CodeContext);
 
   if (!codeContext) {
     throw new Error("CodeContext not found");
   }
-  const { code } = codeContext;
-  const { openAlert, closeAlert, CustomAlert } = useCustomAlert();
-  const mutation = useMutation({
-    mutationFn: runCode,
-    async onSuccess(data) {
-      setPreprocessedCodes([]);
-      setCodeFlowLength(0);
-      setStepIdx(0);
-      setConsoleList([data.result.output]);
-      setHighlightLines([]);
-    },
-    onError(error) {
-      console.error(error);
-
-      if (error.message === "데이터 형식이 올바르지 않습니다") {
-        return;
-      } else if ((error as any).code === "CA-400006" || (error as any).code === "CA-400999") {
-        return;
-      } else if ((error as any).code === "CA-400002") {
-        // 잘못된 문법 에러처리
-        const linNumber = Number((error as any).result.lineNumber);
-        const errorMessage = (error as any).result.errorMessage;
-        setErrorLine({ lineNumber: linNumber, message: errorMessage });
-        setConsoleList([errorMessage]);
-        setPreprocessedCodes([]);
-        return;
-      } else if ((error as any).code == "CA-400007") {
-        alert("코드의 실행 횟수가 너무 많습니다.");
-        return;
-      }
-      setConsoleList([]);
-    },
-  });
-  const handleRunCode = () => {
-    openAlert('클릭했습니다')
-    mutation.mutate({ code, inputData });
-  };
 
   return (
     <Fragment>
@@ -82,15 +31,6 @@ const LeftSection = () => {
         <div className={styles["top-bar"]}>
           <p className={styles["view-section-title"]}>코드 작성</p>
           <div className="flex items-center gap-4">
-            <button
-              type="button"
-              className={`${styles["playcode-btn"]} ${mutation.isPending ? styles["playcode-btn-loading"] : ""}`}
-              onClick={handleRunCode}
-              disabled={mutation.isPending} // 로딩 중에는 버튼 비활성화
-            >
-              <img src="/image/icon_play_w.svg" alt="" />
-              결과보기
-            </button>
             <Dropdown />
           </div>
         </div>
